pyproject.toml
[project]
name = "oracle-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "playwright>=1.54.0",
    "playwright-stealth>=2.0.0",
    "pydantic-ai-slim[google]>=0.5.1",
]
